backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
requests==2.31.0

